+++ src/zotero_project_manager/__init__.py
@@ -1,3 +1,3 @@
 """Zotero Project Manager package."""
 
-__version__ = "1.1.0.dev4"
+__version__ = "1.1.0.dev5"